Split yum command implementations to separate class
